eel2-arm: more stable NSEEL_code_execute(..) calls
prevent a crash due to the link register (LR) being updated
incorrectly from the stack frame. this needs to be addressed
correctly later on in GLUE_FUNC_LEAVE[n].

while the code executes, nothing happens yet for the basic
assign operator in test.cpp (ret = ...)

the gdb and redhat-insight are tremendous help for debugging.

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>