added min, max, invsqrt approximation
- the invsqrt version is for single precision since the compiler
does not return correct values with soft-float doubles (?)
- it seems that moving the stack pointer around might not be a
good idea in general on arm or atleast the defined offsets cause
unexpected behavior
- using localy defined functions (see nseel_mul), could be
method that leaves the option to choose the soft/hard floating
point implementation from the compiler. of course with
the expense of an extra branching. also these has to be defined
in _flist0

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>