added implementation scheme for the bitwise operators
- without fast floating point instruction to truncate doubles,
we relly on the compiler to do the work for us. once truncated we
cast them to 64bit integers and can perform bitwise operations
with them. the alternative is to make the trucation magic ourself -
working on a couple of 32bit registers that hold the double
- the mentioned above use an added local definition of floor which is
faster than the one in libc / soft-float
- added prologue/epilogue for the << and >> operators

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>