NSEEL_DECLARE_LOCAL_BOP_FP2: cast to 32bit integers instead of 64bit
the NSEEL_DECLARE_LOCAL_BOP_FP2 macro expands to a function that
takes a couple of doubles, casts them to [*]integer and then performs
either a bit-wise operation, a bit-shift or modulo. the result
is casted back to a double.

a possible optimization is to write fast local, arch. independent
version of functions such as: double2int() and int2double().

[*] to make the generated output values the same as the x86 and
ppc versions, we should cast to 32bit integer instead of 64bit.

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>