merged the eel2-arm port into the official WDL (a8a666ba913d0)
a lot of refactoring has been done in EEL2 over the last month,
so this port is currently broken. i was able to make it compile at least,
but will have to look into making the necessary changes.

the structuring and naming looks much nicer now, but i haven't tested
any of the new functionality at this point.

this is a new branch and a manual merge, since there were way to
many merge conflicts. :(

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>